refactor: [renderer] N9: relocate default style values to styles/defaults.ts (values unchanged, exports intact)
