+++ src/docIndex.ts
@@ -16,7 +16,7 @@ import type { GraphRenderer, AbstractSelectionBox } from './GraphRenderer'
 import type { Simulation } from './Simulation'
 import type { Notifier, NotificationLevel } from './ui/Notifier'
 import type { TreeLayoutAlgorithm } from './plugins/layout/Tree'
-import { defaultNodeStyle, defaultEdgeStyle, defaultLabelStyle, defaultMarkerStyleMap } from './renderers/svg/GraphSvgRenderer'
+import { defaultNodeStyle, defaultEdgeStyle, defaultLabelStyle, defaultMarkerStyleMap } from './styles/defaults'
 import type { DeepPartial } from './utils/utils'
 import type { UIElement } from './ui/UIManager'
 import { UIComponent } from './ui/UIComponent'
