+++ src/renderers/svg/NodeDrawer.ts
@@ -3,7 +3,8 @@ import { transition as d3Transition } from 'd3-transition'
 import { Node } from '../../Node'
 import { Edge } from '../../Edge'
 import type { Graph } from '../../Graph'
-import { GraphSvgRenderer, defaultLabelStyle } from './GraphSvgRenderer'
+import { GraphSvgRenderer } from './GraphSvgRenderer'
+import { defaultLabelStyle } from '../../styles/defaults'
 import { resolveIcon, tryResolveNumber, tryResolveString } from '../../utils/Getters'
 import type { CustomNodeShape, GraphRendererOptions, ImageFit, NodeShape, NodeStyle } from '../../interfaces/RendererOptions'
 import { ClusterDrawer } from './ClusterDrawer'
